Bind traffic to one address, and give metadata requests owners
Two things drove this. A host that routes through a VPN had no way to
make the engine stay on that interface, and a magnet's metadata fetch
was scheduled per piece, so one silent peer could hold a piece's only
timer while capable peers sat idle.

ConnectionSettings.BindAddress now covers inbound TCP/UDP, LSD, outbound
TCP/uTP, SOCKS control and relay sockets, HTTP and UDP trackers, web
seeds, and exact-source magnet fetches. A configured address never falls
back to Any - socket creation fails instead, which is what makes it
usable as a kill switch. Port mapping is disabled while it is set,
because gateway discovery opens its own interface-selected sockets and
cannot honour the guarantee.

Metadata requests are now peer-owned: each eligible peer independently
picks a least-requested missing piece, and timeouts, rejects and
disconnects are tracked per peer and piece. The pipeline still caps
distinct pieces in flight and MetadataRequestRedundancy caps simultaneous
owners of one piece.

A running magnet also keeps its established sockets across metadata
initialization. Bitfield and HAVE state that arrived while the piece
count was unknown is retained, resized and replayed; anything missing or
inconsistent falls back to close-and-rediscover.

Also here: direct trackers announce independently over IPv4 and IPv6 and
one family failing no longer discards the other's peers; bandwidth limits
and reported rates are signed 64-bit throughout, with negative limits
rejected and quota arithmetic saturating; end-game requests are capped at
four copies per block; and PeerDisconnectedAlert and
MetadataDownloadStalledAlert report what the current-peer snapshot cannot.

+++ src/PeerSharp/Core/EngineStats.cs
@@ -11,8 +11,8 @@ namespace PeerSharp.Core;
 /// <param name="ActiveTorrents">The number of torrents that are currently active.</param>
 /// <param name="TotalPeers">The total number of connected peers across all torrents.</param>
 public sealed record EngineStats(
-    int DownloadSpeed = 0,
-    int UploadSpeed = 0,
+    long DownloadSpeed = 0,
+    long UploadSpeed = 0,
     long TotalDownloaded = 0,
     long TotalUploaded = 0,
     int TorrentCount = 0,
