+++ tests/PeerSharp.Tests/Core/Peers/ProtocolEncryptionTests.cs
@@ -57,10 +57,10 @@ public class ProtocolEncryptionTests
         public IReadOnlyList<int> StreamableFileIndices => Array.Empty<int>();
         public ITorrentEvents? Events => null;
         public DownloadStrategy DownloadStrategy { get => DownloadStrategy.RarestFirst; set { } }
-        public int DownloadLimitBytesPerSecond { get => 0; set { } }
+        public long DownloadLimitBytesPerSecond { get => 0; set { } }
         public int DiskReadLimitBytesPerSecond { get => 0; set { } }
         public int DiskWriteLimitBytesPerSecond { get => 0; set { } }
-        public int UploadLimitBytesPerSecond { get => 0; set { } }
+        public long UploadLimitBytesPerSecond { get => 0; set { } }
         public int QueuePriority { get => 0; set { } }
         public bool QueueAutoStart { get => false; set { } }
         public float? RatioLimit { get => null; set { } }
@@ -456,7 +456,6 @@ public class ProtocolEncryptionTests
 
     #endregion
 }
-
